refactor: merge poi mods
