+++ graph-gateway/tests/it_indexings_blocklist.rs
@@ -4,8 +4,8 @@ use alloy_primitives::Address;
 use thegraph::types::DeploymentId;
 use tokio::time::timeout;
 
+use graph_gateway::indexers::public_poi::{ProofOfIndexing, ProofOfIndexingInfo};
 use graph_gateway::indexings_blocklist::check_indexer_pois;
-use graph_gateway::poi::{ProofOfIndexing, ProofOfIndexingInfo};
 
 /// Test utility function to create a valid `ProofOfIndexingInfo` with an zeros POI.
 fn zero_poi() -> ProofOfIndexing {
